fix: break IModelHandler <-> ModelHandler type-level import cycle
IModelHandler.ts imported ModelHandler.ts (for its Pick) while
ModelHandler.ts imported several shared request/result/tool-call types
from IModelHandler.ts, creating a type-only circular dependency flagged
by review.

Move the plain data contracts that don't reference ModelHandler
(CreateResponseOptions, CreateResponseResult, ExtractResponseResult,
StopConditionsResult, TokenCountOptions, TokenValidationResult, and the
per-provider ToolCall types plus SdkToolCall) into a new
types/ModelHandlerContracts.ts module. ModelHandler.ts now imports them
from there instead of from IModelHandler.ts. IModelHandler.ts re-exports
them for existing consumers, so no other call sites need to change.
