+++ src/agent/modelHandlers/types/IModelHandler.ts
@@ -1,176 +1,33 @@
 // Local imports - agent components
 import type { AgentWorkspaceState } from '@agent/core/state/AgentWorkspaceState';
-import type { ProviderUsage } from '@agent/core/usage/ResponseUsage';
-import type { ToolDefinition } from '@model';
 import type {
   ToolFileAttachment,
   ToolResult,
 } from '@shared/schemas/toolResult';
 import type { ModelHandler } from '../ModelHandler';
 import type { ProviderMessage } from './ProviderMessage';
-import type { ProviderStopReason } from './StopReasonTypes';
-import type {
-  ChatCompletionMessageFunctionToolCall,
-  ChatCompletionMessageToolCall,
-} from 'openai/resources/chat/completions';
-import type { ResponseFunctionToolCallItem } from 'openai/resources/responses/responses';
-import type { FunctionCall } from '@google/genai';
-import type { ToolUseBlock } from '@anthropic-ai/sdk/resources/messages';
-import type { ChatToolCall as ORChatToolCall } from '@openrouter/sdk/models';
+import type { SdkToolCall } from './ModelHandlerContracts';
 
-/**
- * Options for token counting - all fields optional, handlers use what they need.
- * @template C Provider-specific client type
- */
-export interface TokenCountOptions<C = unknown> {
-  /** Pre-authenticated client (avoids re-auth) */
-  client?: C;
-  /** System prompt to include in count */
-  systemPrompt?: string;
-  /** Tool definitions to include in count */
-  tools?: unknown[];
-  /** For cancellation */
-  signal?: AbortSignal;
-}
-
-/**
- * Result from validating token limits against context window.
- */
-export interface TokenValidationResult {
-  /** Adjusted max tokens to fit within context window */
-  adjustedMaxTokens: number;
-  /** Input token count used for validation */
-  inputTokens?: number;
-  /** Percentage of context window utilized */
-  utilizationPercent?: number;
-}
-
-/**
- * Options for creating a model response.
- * @template M - Provider-specific message type
- */
-export interface CreateResponseOptions<
-  M extends ProviderMessage = ProviderMessage,
-  C = unknown,
-> {
-  /** Provider client instance */
-  client: C;
-  /** Conversation messages */
-  messages: M[];
-  /** Sampling temperature (0-1) */
-  temperature: number;
-  /** Optional system prompt */
-  systemPrompt?: string;
-  /** Optional stop sequence */
-  endTag?: string;
-  /** Optional abort signal for cancellation */
-  signal?: AbortSignal;
-  /** Optional tool definitions for function calling */
-  tools?: ToolDefinition[];
-}
-
-/**
- * Result from createResponse, optionally including updated messages
- * after compaction or other transformations.
- *
- * @template Resp - Provider-specific response type
- * @template M - Provider-specific message type
- */
-export interface CreateResponseResult<
-  Resp,
-  M extends ProviderMessage = ProviderMessage,
-> {
-  /** The provider response */
-  response: Resp;
-  /**
-   * If messages were transformed (e.g., compaction), the new array.
-   * Undefined means messages were not modified - caller keeps original.
-   */
-  updatedMessages?: M[];
-}
-
-/**
- * Result from extracting response data from a provider response.
- */
-export interface ExtractResponseResult {
-  /** Extracted response text */
-  text: string;
-  /** Usage/token statistics from the provider */
-  usage: ProviderUsage;
-  /** Reason why the model stopped generating */
-  stopReason: ProviderStopReason;
-}
-
-/**
- * Result from checking stop conditions.
- */
-export interface StopConditionsResult {
-  /** Whether the turn should end */
-  endTurn: boolean;
-  /** Whether generation should stop */
-  shouldStop: boolean;
-}
-
-export type OpenAIToolCall = {
-  provider: 'openai';
-  callId: string;
-  name: string;
-  input: ChatCompletionMessageFunctionToolCall['function']['arguments'];
-  raw: ChatCompletionMessageToolCall;
-};
-
-export type DeepSeekToolCall = {
-  provider: 'deepseek';
-  callId: string;
-  name: string;
-  input: unknown;
-  raw: ChatCompletionMessageToolCall;
-};
-
-export type OpenAIResponseToolCall = {
-  provider: 'openai-response';
-  callId: string;
-  name: string;
-  input: unknown;
-  raw: ResponseFunctionToolCallItem;
-};
-
-export type GoogleToolCall = {
-  provider: 'google';
-  callId: string;
-  name: string;
-  input: FunctionCall['args'];
-  raw: FunctionCall;
-  thoughtSignature?: string;
-};
-
-export type AnthropicToolCall = {
-  provider: 'anthropic';
-  callId: string;
-  name: string;
-  input: ToolUseBlock['input'];
-  raw: ToolUseBlock;
-};
-
-export type OpenRouterToolCall = {
-  provider: 'openrouter';
-  callId: string;
-  name: string;
-  input: unknown;
-  raw: ORChatToolCall;
-};
-
-export type SdkToolCall =
-  | OpenAIToolCall
-  | DeepSeekToolCall
-  | OpenAIResponseToolCall
-  | GoogleToolCall
-  | AnthropicToolCall
-  | OpenRouterToolCall;
-
-// Note: SdkToolCall is a discriminated union on 'provider'.
-// Use `call.provider === 'openai'` directly for type narrowing instead of
-// separate type guard functions.
+// Re-exported for existing consumers of this module. These plain data
+// contracts live in `./ModelHandlerContracts` (not here) because
+// `ModelHandler.ts` needs them too, and importing them from this file would
+// recreate the ModelHandler.ts <-> IModelHandler.ts cycle this module exists
+// to avoid — see the note on `IModelHandler` below.
+export type {
+  TokenCountOptions,
+  TokenValidationResult,
+  CreateResponseOptions,
+  CreateResponseResult,
+  ExtractResponseResult,
+  StopConditionsResult,
+  OpenAIToolCall,
+  DeepSeekToolCall,
+  OpenAIResponseToolCall,
+  GoogleToolCall,
+  AnthropicToolCall,
+  OpenRouterToolCall,
+  SdkToolCall,
+} from './ModelHandlerContracts';
 
 /**
  * Common port implemented by all model handlers.
